تغییرات قسمت مشتری و کارشناس برای نمایش سوالات و قرار دادن نمره برای آنها قسمت نمره هنوز کامل نشده است و نیاز به ارسال نمره تایید شده کارشناس به سمت سرور است

+++ ParsKyanCrm.Domain/Entities/DataFormReport.cs
@@ -11,8 +11,8 @@ namespace ParsKyanCrm.Domain.Entities
         public int DataReportId { get; set; }
         public int RequestId { get; set; }
         public int DataFormAnswerId { get; set; }
-        public int SystemScore { get; set; }
-        public int AnalizeScore { get; set; }
+        public float SystemScore { get; set; }
+        public float AnalizeScore { get; set; }
         public string Description { get; set; }
         public byte IsActive { get; set; }
     }
